Feature/ Add series to link related articles
